Add in Zale, Mordessa, Staff of Earth Spirit, Mordessa's Staff
When the player finishes the last stage of each campaign, they unlock Zale (Light) or Mordessa (Dark), and the Staff of the Earth Spirit (Light) for Druid, Wizard, Zale; and Mordessa's Staff (Dark) for Sorceress, Lich, Mordessa. Quit to menu after finishing the campaign has been removed to allow for this after-game update.